a minimum refactor of optimizer and design region
